Crate collectives_infrablockspace_runtime

Collectives Parachain

This parachain is for collectives that serve the Polkadot network. Each collective is defined by a specialized (possibly instanced) pallet.

Governance

As a common good parachain, Collectives defers its governance (namely, its Root origin), to its Relay Chain parent, Polkadot.

Collator Selection

Collectives uses pallet-collator-selection, a simple first-come-first-served registration system where collators can reserve a small bond to join the block producer set. There is no slashing. Collective members are generally expected to run collators.

Re-exports

pub use parachains_common as common;

Modules

api

constants

impls

xcm_config

Structs

AllianceMotionDuration

AllianceRetirementPeriod

AllyDeposit

AnnouncementDepositBase

AnnouncementDepositFactor

DepositBase

DepositFactor

ExecutiveBody

ExistentialDeposit

GenesisConfig

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotId

ProxyDepositBase

ProxyDepositFactor

RelayTreasuryAccId

ReservedDmpWeight

ReservedXcmpWeight

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeBlockLength

RuntimeBlockWeights

RuntimeOrigin

The runtime origin type representing the origin of a call.

SessionKeys

SessionLength

SlashedImbalanceAccId

TransactionByteFee

Relay Chain TransactionByteFee / 10

Version

Enums

OriginCaller

ProxyType

The type used to represent the kinds of proxying allowed.

RuntimeCall

RuntimeEvent

Constants

ALLIANCE_MAX_MEMBERS

ALLIANCE_MAX_PROPOSALS

ALLIANCE_MOTION_DURATION

MAX_ALLIES

MAX_FELLOWS

OFFSET

PERIOD

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Traits

BuildStorage

Complex storage builder stuff.

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Definitions

Address

The address format for describing accounts.

AllPalletsDeprecated

All pallets included in the runtime as a nested tuple of types.

AllPalletsReversedWithSystemFirstDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. With the system pallet first.

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllPalletsWithoutSystemReversedDeprecated

All pallets included in the runtime as a nested tuple of types in reversed order. Excludes the System pallet.

Alliance

AllianceConfig

AllianceMotion

AllianceMotionConfig

Aura

AuraConfig

AuraExt

AuraExtConfig

Authorship

Balances

BalancesConfig

Block

Block type as expected by this runtime.

BlockId

BlockId type as expected by this runtime.

CheckedExtrinsic

Extrinsic type that has already been checked.

CollatorSelection

CollatorSelectionConfig

CollatorSelectionUpdateOrigin

We allow root and the Relay Chain council to execute privileged collator selection operations.

CumulusXcm

DmpQueue

Executive

Executive: handles dispatch to the various modules.

Multisig

ParachainInfo

ParachainInfoConfig

ParachainSystem

ParachainSystemConfig

PolkadotXcm

PolkadotXcmConfig

Proxy

RootOrAllianceTwoThirdsMajority

Privileged origin that represents Root or more than two thirds of the Alliance.

RootOrExecutiveSimpleMajority

Privileged origin that represents Root or the majority of the Relay Chain Council.

Session

SessionConfig

SignedBlock

A Block signed with a Justification

SignedExtra

The SignedExtension to the basic transaction logic.

System

SystemConfig

Timestamp

TransactionPayment

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility

XcmpQueue